% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_start_datetime_minidot.R
\name{detectDrift}
\alias{detectDrift}
\title{Correct starting datetimes in MiniDot files}
\usage{
detectDrift(
  folderpath,
  reference_file = NULL,
  test_start,
  test_end,
  tz = 0,
  threshold_drift = 0.4,
  patterns2searchfor = NULL,
  patterns2exclude = c("00_drift_results", "@",
    "pres_air|pres_eau|pres.txt|pres_cote|pression|Pres.csv", "aff.txt", "status.txt",
    "lire.txt", "Sorties.txt", "aff_quinson", "allos CR.txt", "00_ALIRE.txt"),
  writeOutput = TRUE,
  out_dir = NULL
)
}
\arguments{
\item{file}{Path to the MiniDot file to correct.}

\item{firstdatetime}{The true datetime of the first observation (used when
correcting sequentially).}

\item{lagdatetime}{A character vector of length two: the sensor-recorded datetime
and the corresponding correct datetime.}

\item{nskip}{Number of header rows to skip before the data section. Default is 9,
consistent with the MiniDot file template.}

\item{append}{String to append to the output file name to avoid overwriting the
original. Default is "_datetimecorrected".}

\item{overwrite}{Logical. If \code{FALSE} (default), the function will stop if the
output file already exists. If \code{TRUE}, the existing file will be overwritten.}
}
\description{
MiniDot compiled files often contain errors in their recorded datetimes.
This function provides two ways to correct them:
\enumerate{
\item \strong{Using \code{firstdatetime}:} specify the true datetime of the first observation.
The function will then incrementally correct all subsequent rows based on the
expected sampling interval.
\item \strong{Using \code{lagdatetime}:} supply a pair of datetimes, where the first is the
incorrect value recorded by the sensor, and the second is the correct value
(e.g., as determined by the computer during download). The function applies
this offset to all records in the file.
}

The function preserves the original file structure and header rows, and writes
out a corrected version with a modified file name.
}
\examples{
correct_start_datetime_minidot(file = "/Users/robruel/Desktop/Travail/ScriptR/RNT/Input/Data/SCL06/sonde/20250724/SCL06_00500_O2.TXT",
    lagdatetime = c("2024-04-25 10:06:47", "2025-07-24 12:05:47"), overwrite = TRUE)
correct_start_datetime_minidot(file = "/Users/robruel/Desktop/Travail/ScriptR/RNT/Input/Data/SCL06/sonde/20250724/SCL06_00070_O2.TXT",
lagdatetime = c("2023-06-30 22:18:17", "2025-07-24 11:55:17"),
overwrite = TRUE)
}
\author{
Rosalie Bruel, OFB, Pôle R&D ECLA
}
