% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAUD_find.date.first.mixing.R
\name{find.date.first.mixing}
\alias{find.date.first.mixing}
\title{Find lake spring mixing using temperature measured in the epi- and hypolimnion}
\usage{
find.date.first.mixing(
  dt,
  DateTime.col,
  Temp.Surf.col,
  Temp.Deep.col,
  YearBrass,
  start.date,
  end.date,
  grouping_var = NULL,
  optioncalcul,
  duration2test = NULL,
  prop.data.available,
  thresh,
  duration,
  error.tolerated,
  which.val.to.substract = 1,
  absolutevalue = FALSE
)
}
\arguments{
\item{dt}{dataset}

\item{DateTime.col}{name of the column with the date time information or integer of column number.}

\item{Temp.Surf.col}{name of the column with the surface-water temperature data (or integer of column number).}

\item{Temp.Deep.col}{name of the column with the deep-water temperature data (or integer of column number).}

\item{YearBrass}{Year being looked at.}

\item{start.date}{Starting date (julian date, 1-366) to crop data.}

\item{end.date}{End date (julian date, 1-366) to crop data.}

\item{grouping_var}{Additional grouping variables, such as SiteID, depth_m, sensor, etc.  Must be columns names that exist in dt.}

\item{optioncalcul}{Options are c("belowthresh", "inversion", "minsd"). Different ways of calculating the date of mixing: with a value below a certain threshold abs(Val1-Val2)<thresh with thresh = 0.5 for instance, using the moment of inversion (Val1-Val2)>thresh with thresh = 0 typically, or the minimum sd between the two values, assuming that low sd = the moment when the two values are the closest = the moment of mixing. Options are c("belowthresh", "inversion", "minsd").}

\item{prop.data.available}{A proportion (0-1) of temperature data that needs to be available to carry on and trust the result}

\item{thresh}{Threshold: the date we will return must be below that threshold.}

\item{duration}{Duration the difference must last for. Default = 1: only one observation}

\item{error.tolerated}{Proportion (0-1) of the duration that can be above of the selected threshold. Default = 1, ALL the values must be within the difference.}

\item{which.val.to.substract}{Argument for the function find.date.value.above.threshold that gives the direction of the difference. Default = 1, if you want to do Temp.Surf.col-Temp.Deep.col If which.val.to.substract = 2, then the function will do Temp.Deep.col - Temp.Surf.col}

\item{absolutevalue}{Default = FALSE. If TRUE, will find the first value that is above the threshold, using the absolute difference between Val1 and Val2.}
}
\description{
Take in a dataset at hourly time-interval with shallow and deep depth and find the likely mixing date. This function was developped for mountain lakes and cannot necessarily be reapplied directly to other lakes.
}
\author{
Rosalie Bruel
}
