% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAUD_find.date.min.sd.between.2values.R
\name{find.date.min.sd.between.2values}
\alias{find.date.min.sd.between.2values}
\title{Return the date with the minimum SD between 2 values}
\usage{
find.date.min.sd.between.2values(
  dt,
  DateTime.col,
  Value1.col,
  Value2.col,
  thresh = NULL
)
}
\arguments{
\item{dt}{dataset}

\item{DateTime.col}{name of the column with the date time information or integer of column number. Also takes in date. Format must be "\%Y-\%m-\%d \%H:\%M:\%S" or "\%Y-\%m-\%d"}

\item{Value1.col}{name of the column with the 1st value data (or integer of column number).}

\item{Value2.col}{name of the column with the 2nd value data (or integer of column number).}

\item{thresh}{Threshold: the date we will return should have a sd between val 1 and 2 below that threshold (Useful so that no value is return if sd too high. Default = NA}
}
\description{
Return the date with the minimum sd
}
\author{
Rosalie Bruel
}
