% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAUD_find.date.value.above.threshold.R
\name{find.date.value.above.threshold}
\alias{find.date.value.above.threshold}
\title{Return the first date with an inversion between 2 value (when Val1>Val2).}
\usage{
find.date.value.above.threshold(
  dt,
  DateTime.col,
  Value1.col,
  Value2.col,
  which.val.to.substract = 1,
  thresh = 0,
  difference = 0,
  duration,
  error.tolerated = 1,
  absolutevalue = FALSE
)
}
\arguments{
\item{dt}{dataset}

\item{DateTime.col}{name of the column with the date time information or integer of column number. Also takes in date. Format must be "\%Y-\%m-\%d \%H:\%M:\%S" or "\%Y-\%m-\%d"}

\item{Value1.col}{name of the column with the 1st value data (or integer of column number).}

\item{Value2.col}{name of the column with the 2nd value data (or integer of column number).}

\item{which.val.to.substract}{Default = 1, if you want to do Value1.col-Value2.col. If which.val.to.substract = 2, then the function will do Value2.col - Value1.col.}

\item{thresh}{Threshold: the date we will return must be below that threshold. Default = 0 to find the inversion.}

\item{difference}{A value to add to 'thresh': in some cases, measurements are a bit imprecise, and it may be useful to confirm we went above that value. Typically with soil or lakes, we can specify for example 2: after 15 days, we expect the value to have gone above that additional value if we really want to capture the process. Default = 0.}

\item{duration}{Duration the difference must last for. Default = 1: only one observation}

\item{error.tolerated}{Proportion (0-1) of the duration that can be above of the selected threshold. Default = 1, ALL the values must be within the difference.}

\item{absolutevalue}{Default = FALSE. If TRUE, will find the first value that is above the threshold, using the absolute difference between Val1 and Val2.}
}
\description{
Return the first date with an inversion between 2 value (when Val1>Val2). The function allows to decide whether we want it to last for several a period.
}
\author{
Rosalie Bruel
}
