% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAUD_find.date.value.below.thresh.R
\name{find.date.value.below.thresh}
\alias{find.date.value.below.thresh}
\title{Return the first date with a value below a certain threshold.}
\usage{
find.date.value.below.thresh(
  dt,
  DateTime.col,
  Value.col,
  thresh,
  duration,
  error.tolerated = 1
)
}
\arguments{
\item{dt}{dataset}

\item{DateTime.col}{name of the column with the date time information or integer of column number. Also takes in date. Format must be "\%Y-\%m-\%d \%H:\%M:\%S" or "\%Y-\%m-\%d"}

\item{Value.col}{name of the column with the value data (or integer of column number).}

\item{thresh}{Threshold: the date we will return must be below that threshold. The values are turned as absolute, so if thresh = 0.3, the date will have a corresponding value between -0.3 and 0.3.}

\item{duration}{Duration the difference must last for. Default = 1: only one observation}

\item{error.tolerated}{Proportion (0-1) of the duration that can be above of the selected threshold. Default = 1, ALL the values must be within the difference.}
}
\description{
Return the first date with a value below a threshold. The function allows to decide whether we want that period to last for several iterations.
}
\author{
Rosalie Bruel
}
