% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasqas-package.R
\docType{package}
\name{rasqas-package}
\alias{rasqas}
\alias{rasqas-package}
\title{rasqas: Read, Assemble, Standardize, Qualify and Analyze data from autonomous Sensors}
\description{
Functions to read, assemble, and standardize high-frequency data collected by autonomous environmental sensors. The package currently provides functions to ingest heterogeneous text and CSV files produced by multiple sensor manufacturers, automatically detect date-time formats, clean and harmonize measurements, and assess sensor drift compared to a reference sensor. Future developments will extend the package with additional methods for data qualification, analysis, and visualization of sensor time series.
}
\author{
\strong{Maintainer}: Rosalie Bruel \email{rosalie.bruel@ofb.gouv.fr} (\href{https://orcid.org/0000-0001-8351-7539}{ORCID})

}
\keyword{internal}
